requirementsNoOpenCv.txt
paho-mqtt==1.5.1
Pillow>=6.2.0
aiohttp>=3.8.6